리팩토링 Login, signUp
