+++ src/pages/Home.tsx
@@ -1,12 +1,10 @@
 import React from 'react';
 import Header from '../components/common/Header';
-import Post from '../components/Post';
 
 const Home = () => {
   return (
     <div>
       <Header />
-      <Post />
     </div>
   );
 };
